+++ src/hotspot/share/prims/methodHandles.cpp
@@ -93,7 +93,7 @@ void MethodHandles::generate_adapters() {
   // The adapter entry is required to be aligned to CodeEntryAlignment.
   // So we need additional bytes due to alignment.
   int adapter_num = (int)Interpreter::method_handle_invoke_LAST - (int)Interpreter::method_handle_invoke_FIRST + 1;
-  int max_aligned_bytes = adapter_num * CodeEntryAlignment;
+  int max_aligned_bytes = adapter_num * (int)CodeEntryAlignment;
   _adapter_code = MethodHandlesAdapterBlob::create(adapter_code_size + max_aligned_bytes);
   CodeBuffer code(_adapter_code);
   MethodHandlesAdapterGenerator g(&code);
